Add progress delay for ~2x speedup when progress indicators are on
Progress indicators were updating for every single operation that
was performed. This can be extremely slow because each update calls
fprintf() to update the display. It is even slower via non-local
terminals like SSH.

I have added a delay counter to space out progress updates in this
patch. The result is that performance is almost equal to the
performance obtained by running with progress indicators disabled.
Typical speedup seems to be approximately two times faster.