add reflink support
Add option --reflink to deduplicate identical files via filesystem reflinks.
Reflink is done via the FICLONE ioctl, and allows to share the file
content across multiple inodes. This is similar to hardlink, with two
substantial differences:
- each copy has its own metadata (permissions, timestamps, attributes, etc.)
- on write, the content is copied (CoW) so the changes are not reflected to all other copies

Currently implemented only on Linux, it works with BtrFS and XFS (and ZFS to some extents)